♻️ Refactor base models
